+++ straxen/plugins/events/__init__.py
@@ -43,11 +43,11 @@ from .s2_recon_pos_diff import *
 from . import veto_proximity
 from .veto_proximity import *
 
-from . import event_s2position_mlp
-from .event_s2position_mlp import *
+from . import event_s2_position_mlp
+from .event_s2_position_mlp import *
 
-from . import event_s2position_cnn
-from .event_s2position_cnn import *
+from . import event_s2_position_cnn
+from .event_s2_position_cnn import *
 
-from . import event_s2position_gcn
-from .event_s2position_gcn import *
+from . import event_s2_position_gcn
+from .event_s2_position_gcn import *
