+++ straxen/plugins/events/event_s2_position_cnn.py
@@ -7,21 +7,21 @@ export, __all__ = strax.exporter()
 
 
 @export
-class EventS2PositionMLP(EventS2PositionBase):
+class EventS2PositionCNN(EventS2PositionBase):
     """
-    MLP neural net for S2 position reconstruction at event level
+    CNN for position S2 reconstruction at event level
     """
-    algorithm = "mlp"
-    provides = "event_s2position_mlp"
+    algorithm = "cnn"
+    provides = "event_s2_position_cnn"
 
-    tf_event_model_mlp = straxen.URLConfig(
+    tf_event_model_cnn = straxen.URLConfig(
         default=f'tf://'
                 f'resource://'
                 f'cmt://{algorithm}_model'
                 f'?version=ONLINE'
                 f'&run_id=plugin.run_id'
                 f'&fmt=abs_path',
-        help='MLP model. Should be opened using the "tf" descriptor. '
+        help='CNN model. Should be opened using the "tf" descriptor. '
              'Set to "None" to skip computation',
         cache=3,
     )
